make master up tp date

+++ common/makeSampleList.py
@@ -14,26 +14,38 @@ import EventProducer.common.utils as ut
 class makeSampleList():
 
 #__________________________________________________________
-    def __init__(self, para, version):
+    def __init__(self, para, version, detector):
         self.para = para
-        self.heppyList = self.para.heppyList.replace('VERSION',version)
-        self.procList  = self.para.procList.replace('VERSION',version)
+        self.heppyList = self.para.heppyList.replace('VERSION',version).replace('DETECTOR',detector)
+        self.procList  = self.para.procList.replace('VERSION',version).replace('DETECTOR',detector)
         self.version   = version
+        self.detector  = detector
 #______________________________________________________________________________________________________
     def addEntry(self, process, yaml_lhe, yaml_reco, xsec, kf, heppyFile, procDict,proc_param=''):
         processhad=process
         if 'mgp8_' in process:
             processhad=process.replace('mgp8_','mg_')
+        #elif 'wzp6_' in process:
+        #    processhad=process.replace('wzp6_','wz_')
+        elif 'wzp8_' in process:
+            processhad=process.replace('wzp8_','wz_')
+        elif 'pwp8_' in process:
+            processhad=process.replace('pwp8_','pw_')
+        elif 'kkmcp8_' in process:
+            processhad=process.replace('kkmcp8_','kkmc_')
+
+            
         if  proc_param!='':
             processhad=proc_param.replace('mgp8_','mg_')
+            processhad=proc_param.replace('kkmcp8_','kkmc_')
 
         yaml_lhe=yaml_lhe+'/'+processhad+'/merge.yaml'
-        print 'lhe yaml    ',yaml_lhe
-        print 'reco yaml    ',yaml_reco
+        print ('lhe yaml    ',yaml_lhe)
+        print ('reco yaml    ',yaml_reco)
 
         if not ut.file_exist(yaml_lhe): 
-            print 'no merged file lhe for process %s continue'%process
-            sys.exit(3)
+            print ('no merged file lhe for process %s continue'%process)
+            #sys.exit(3)
             return 1.0
 
         
@@ -62,9 +74,12 @@ class makeSampleList():
             except yaml.YAMLError as exc:
                 print(exc)
 
+        print (yaml_reco)
         nmatched+= int(yreco['merge']['nevents'])
-        nweights+= float(yreco['merge']['sumofweights'])
-
+        try:
+            nweights+= float(yreco['merge']['sumofweights'])
+        except KeyError as e:
+            print ('there is a KeyError  :  ',e)
         for f in ylhe['merge']['outfiles']:
            
             if any(f[0].replace('.lhe.gz','') in s[0] for s in yreco['merge']['outfiles']):
@@ -73,25 +88,33 @@ class makeSampleList():
                     heppyFile.write("           '/{}/{}',\n".format(yreco['merge']['outdir'],f[0].replace('.lhe.gz','.root')))
                 else:
                     heppyFile.write("           '/{}{}',\n".format(yreco['merge']['outdir'],f[0].replace('.lhe.gz','.root')))
+                    
+            if any(f[0].replace('.stdhep.gz','') in s[0] for s in yreco['merge']['outfiles']):
+                nlhe+=int(f[1])
+                if yreco['merge']['outdir'][-1]!='/':
+                    heppyFile.write("           '/{}/{}',\n".format(yreco['merge']['outdir'],f[0].replace('.stdhep.gz','.root')))
+                else:
+                    heppyFile.write("           '/{}{}',\n".format(yreco['merge']['outdir'],f[0].replace('.stdhep.gz','.root')))
 
+                    
         heppyFile.write(']\n')
         heppyFile.write(')\n')
         heppyFile.write('\n')
 
         # skip process if do not find corresponding lhes
         if nlhe == 0:
-            print 'did not find any LHE event for process', process
+            print ('did not find any LHE event for process', process)
             return matchingEff
        
         if nmatched == 0:
-            print 'did not find any FCCSW event for process', process
+            print ('did not find any FCCSW event for process', process)
             return matchingEff
 
         # compute matching efficiency
         matchingEff = round(float(nmatched)/nlhe, 3)
         if nweights==0: nweights=nmatched
         entry = '   "{}": {{"numberOfEvents": {}, "sumOfWeights": {}, "crossSection": {}, "kfactor": {}, "matchingEfficiency": {}}},\n'.format(process, nmatched, nweights, xsec, kf, matchingEff)
-        print 'N: {}, Nw:{}, xsec: {} , kf: {} pb, eff: {}'.format(nmatched, nweights, xsec, kf, matchingEff)
+        print ('N: {}, Nw:{}, xsec: {} , kf: {} pb, eff: {}'.format(nmatched, nweights, xsec, kf, matchingEff))
 
         procDict.write(entry)
 
@@ -127,24 +150,24 @@ class makeSampleList():
        heppyFile.write('\n')
        
        if nmatched == 0:
-           print 'did not find any FCCSW event for process', process
+           print ('did not find any FCCSW event for process', process)
            return matchingEff
 
        if nweights==0: nweights=nmatched
        entry = '   "{}": {{"numberOfEvents": {}, "sumOfWeights": {}, "crossSection": {}, "kfactor": {}, "matchingEfficiency": {}}},\n'.format(process, nmatched, nweights, xsec, kf, matchingEff)
-       print 'N: {}, Nw:{}, xsec: {} , kf: {} pb, eff: {}'.format(nmatched, nweights, xsec, kf, matchingEff)
+       print ('N: {}, Nw:{}, xsec: {} , kf: {} pb, eff: {}'.format(nmatched, nweights, xsec, kf, matchingEff))
+       print ('entry : ',entry)
        procDict.write(entry)
        return matchingEff
    
 #_______________________________________________________________________________________________________
     def makelist(self):
 
-        yamldir_lhe=self.para.yamldir+'lhe/'
-        yamldir_reco=self.para.yamldir+self.version+'/'
+        yamldir_reco=self.para.yamldir+self.version+'/'+self.detector+'/'
 
         nmatched = 0
         nlhe = 0
-
+        tmpexist=False
         # write header for heppy file
         procDict = open('tmp.json', 'w')
         procDict.write('{\n')
@@ -164,21 +187,32 @@ class makeSampleList():
         for l in ldir:
             processhad=None
             process=l
-
+            if 'mgp8_' in process or 'pwp8_' in process or 'kkmcp8_' in process:
+                yamldir_lhe=self.para.yamldir+'lhe/'
+            elif 'wzp6_' in process or 'wzp8_' in process or 'wz_' in process :
+                yamldir_lhe=self.para.yamldir+'stdhep/'
+                
             yaml_reco=yamldir_reco+'/'+l+'/merge.yaml'
             if not ut.file_exist(yaml_reco): 
-                print 'no merged yaml for process %s continue'%l
+                print ('no merged yaml for process %s continue'%l)
                 continue
 
             #if process!='mgp8_pp_z0123j_4f_HT_5000_100000':continue
-            print ''
-            print '------ process: ', process, '-------------'
-            print ''
+            print ('')
+            print ('------ process: ', process, '-------------')
+            print ('')
             
             if 'mgp8_' in process:
                 processhad=process.replace('mgp8_','mg_')
             elif 'pwp8_' in process:
                 processhad=process.replace('pwp8_','pw_')
+            #elif 'wzp6_' in process:
+            #    processhad=process.replace('wzp6_','wz_')
+            elif 'wzp8_' in process:
+                processhad=process.replace('wzp8_','wz_')
+            elif 'kkmcp8_' in process:
+                processhad=process.replace('kkmcp8_', 'kkmc_')
+                
             else: processhad=process
             # maybe this was a decayed process, so it cannot be found as such in in the param file
             br = 1.0
@@ -189,17 +223,17 @@ class makeSampleList():
                     br = self.para.branching_ratios[dec]
                     decay = dec
             if decay != '':
-                print 'decay---------- '
+                print ('decay---------- ')
                 decstr = '_{}'.format(decay)
                 proc_param = processhad.replace(decstr,'')
-                print '--------------  ',decstr,'  --  ',proc_param
+                print ('--------------  ',decstr,'  --  ',proc_param)
                 
                 try: 
                    xsec = float(self.para.gridpacklist[proc_param][3])*br
                    kf = float(self.para.gridpacklist[proc_param][4])
                    matchingEff = self.addEntry(process, yamldir_lhe, yaml_reco, xsec, kf, heppyFile, procDict,proc_param)
                 except KeyError:
-                   print 'process {} does not exist in the list'.format(process)
+                   print ('process {} does not exist in the list'.format(process))
 
             elif process in self.para.pythialist:
                 xsec = float(self.para.pythialist[process][3])
@@ -208,11 +242,11 @@ class makeSampleList():
  
 
             elif processhad not in self.para.gridpacklist:
-                print 'process :', processhad, 'not found in %s --> skipping process'%self.para.module_name
+                print ('process :', processhad, 'not found in %s --> skipping process'%self.para.module_name)
                 continue
             else: 
-                print 'self.para.gridpacklist[processhad][3]   ',self.para.gridpacklist[processhad][3]
-                print 'self.para.gridpacklist[processhad][4]   ',self.para.gridpacklist[processhad][4]
+                print ('self.para.gridpacklist[processhad][3]   ',self.para.gridpacklist[processhad][3])
+                print ('self.para.gridpacklist[processhad][4]   ',self.para.gridpacklist[processhad][4])
                 xsec = float(self.para.gridpacklist[processhad][3])
                 kf=1
                 if self.para.gridpacklist[processhad][4]!='':
@@ -222,20 +256,20 @@ class makeSampleList():
                 with open(self.para.module_name) as f:
                     lines = f.readlines()
                     isgp=False
-                    for line in xrange(len(lines)):
+                    for line in range(len(lines)):
                         if 'gridpacklist' in str(lines[line]): isgp=True
                         if isgp==False: continue
                         if processhad == lines[line].rsplit(':', 1)[0].replace("'", ""):
                             ll = ast.literal_eval(lines[line].rsplit(':', 1)[1][:-2])
-                            print 'll   : ',ll
-                            print 'line : ',line
-                            print 'phad : ',processhad
-                            print 'meff : ',matchingEff
+                            print ('ll   : ',ll)
+                            print ('line : ',line)
+                            print ('phad : ',processhad)
+                            print ('meff : ',matchingEff)
                             infile[line] = "'{}':['{}','{}','{}','{}','{}','{}'],\n".format(processhad, ll[0],ll[1],ll[2],ll[3],ll[4], matchingEff)
-                            print 'new line ',infile[line]
+                            print ('new line ',infile[line])
                 with open("tmp.py", "w") as f1:
                    f1.writelines(infile)
-
+                   tmpexist=True
         procDict.close()
         # parse param file
 
@@ -251,5 +285,6 @@ class makeSampleList():
         procDict.write('}\n')
         
         # replace existing param.py file
-        os.system("mv tmp.py %s"%self.para.module_name)
+        if tmpexist:
+            os.system("mv tmp.py %s"%self.para.module_name)
         os.system("rm tmp.json")
